Move kv helpers to separate file
