+++ cmd/vault.go
@@ -17,8 +17,6 @@ package cmd
 import (
 	"errors"
 	"fmt"
-	"path"
-	"strings"
 	"text/template"
 
 	"github.com/hashicorp/vault/api"
@@ -51,7 +49,7 @@ func getVaultClient() *api.Client {
 
 		if vaultAuth == vaultRoleAuth {
 			Logger.Debugf(
-				"Getting vault-token using vault-role-id {%s} and vault-secret-id {%s}",
+				"getting vault-token using vault-role-id {%s} and vault-secret-id {%s}",
 				vaultRoleID,
 				vaultSecretID)
 			appRoleLogin := map[string]interface{}{
@@ -63,7 +61,7 @@ func getVaultClient() *api.Client {
 			if resp.Auth == nil {
 				HandleError(
 					errors.New(
-						"Failed to obtain VAULT_TOKEN using vault-role-id and vault-secret-id"))
+						"failed to obtain VAULT_TOKEN using vault-role-id and vault-secret-id"))
 			}
 			vaultToken = resp.Auth.ClientToken
 		}
@@ -112,7 +110,7 @@ func getVaultSecret(path string, key string) string {
 		if !ok {
 			HandleError(
 				fmt.Errorf(
-					"Could not convert vault response [%s][%s] to string",
+					"could not convert vault response [%s][%s] to string",
 					path,
 					key))
 		}
@@ -126,7 +124,7 @@ func getVaultSecret(path string, key string) string {
 		if !ok {
 			HandleError(
 				fmt.Errorf(
-					"Could not convert vault response [%s][%s] to string",
+					"could not convert vault response [%s][%s] to string",
 					path,
 					key))
 		}
@@ -134,84 +132,6 @@ func getVaultSecret(path string, key string) string {
 	}
 
 	return secretStr
-}
-
-
-// copied from github.com/vault/command/kv_helpers
-func addPrefixToVKVPath(p, mountPath, apiPrefix string) string {
-	switch {
-	case p == mountPath, p == strings.TrimSuffix(mountPath, "/"):
-		return path.Join(mountPath, apiPrefix)
-	default:
-		p = strings.TrimPrefix(p, mountPath)
-		return path.Join(mountPath, apiPrefix, p)
-	}
-}
-
-// copied from github.com/vault/command/kv_helpers
-func isKVv2(path string, client *api.Client) (string, bool, error) {
-	mountPath, version, err := kvPreflightVersionRequest(client, path)
-	if err != nil {
-		return "", false, err
-	}
-
-	return mountPath, version == 2, nil
-}
-
-// copied from github.com/vault/command/kv_helpers
-func kvPreflightVersionRequest(client *api.Client, path string) (string, int, error) {
-	// We don't want to use a wrapping call here so save any custom value and
-	// restore after
-	currentWrappingLookupFunc := client.CurrentWrappingLookupFunc()
-	client.SetWrappingLookupFunc(nil)
-	defer client.SetWrappingLookupFunc(currentWrappingLookupFunc)
-	currentOutputCurlString := client.OutputCurlString()
-	client.SetOutputCurlString(false)
-	defer client.SetOutputCurlString(currentOutputCurlString)
-
-	r := client.NewRequest("GET", "/v1/sys/internal/ui/mounts/"+path)
-	resp, err := client.RawRequest(r)
-	if resp != nil {
-		defer resp.Body.Close()
-	}
-	if err != nil {
-		// If we get a 404 we are using an older version of vault, default to
-		// version 1
-		if resp != nil && resp.StatusCode == 404 {
-			return "", 1, nil
-		}
-
-		return "", 0, err
-	}
-
-	secret, err := api.ParseSecret(resp.Body)
-	if err != nil {
-		return "", 0, err
-	}
-	if secret == nil {
-		return "", 0, errors.New("nil response from pre-flight request")
-	}
-	var mountPath string
-	if mountPathRaw, ok := secret.Data["path"]; ok {
-		mountPath = mountPathRaw.(string)
-	}
-	options := secret.Data["options"]
-	if options == nil {
-		return mountPath, 1, nil
-	}
-	versionRaw := options.(map[string]interface{})["version"]
-	if versionRaw == nil {
-		return mountPath, 1, nil
-	}
-	version := versionRaw.(string)
-	switch version {
-	case "", "1":
-		return mountPath, 1, nil
-	case "2":
-		return mountPath, 2, nil
-	}
-
-	return mountPath, 1, nil
 }
 
 // vaultCmdPersistentPreRunE checks required persistent tokens for vaultCmd
