+++ cmd/vault_kv_helpers.go
@@ -1,0 +1,85 @@
+package cmd
+
+import (
+	"errors"
+	"github.com/hashicorp/vault/api"
+	"path"
+	"strings"
+)
+
+// copied from github.com/vault/command/kv_helpers
+func addPrefixToVKVPath(p, mountPath, apiPrefix string) string {
+switch {
+case p == mountPath, p == strings.TrimSuffix(mountPath, "/"):
+return path.Join(mountPath, apiPrefix)
+default:
+p = strings.TrimPrefix(p, mountPath)
+return path.Join(mountPath, apiPrefix, p)
+}
+}
+
+// copied from github.com/vault/command/kv_helpers
+func isKVv2(path string, client *api.Client) (string, bool, error) {
+mountPath, version, err := kvPreflightVersionRequest(client, path)
+if err != nil {
+return "", false, err
+}
+
+return mountPath, version == 2, nil
+}
+
+// copied from github.com/vault/command/kv_helpers
+func kvPreflightVersionRequest(client *api.Client, path string) (string, int, error) {
+// We don't want to use a wrapping call here so save any custom value and
+// restore after
+currentWrappingLookupFunc := client.CurrentWrappingLookupFunc()
+client.SetWrappingLookupFunc(nil)
+defer client.SetWrappingLookupFunc(currentWrappingLookupFunc)
+currentOutputCurlString := client.OutputCurlString()
+client.SetOutputCurlString(false)
+defer client.SetOutputCurlString(currentOutputCurlString)
+
+r := client.NewRequest("GET", "/v1/sys/internal/ui/mounts/"+path)
+resp, err := client.RawRequest(r)
+if resp != nil {
+defer resp.Body.Close()
+}
+if err != nil {
+// If we get a 404 we are using an older version of vault, default to
+// version 1
+if resp != nil && resp.StatusCode == 404 {
+return "", 1, nil
+}
+
+return "", 0, err
+}
+
+secret, err := api.ParseSecret(resp.Body)
+if err != nil {
+return "", 0, err
+}
+if secret == nil {
+return "", 0, errors.New("nil response from pre-flight request")
+}
+var mountPath string
+if mountPathRaw, ok := secret.Data["path"]; ok {
+mountPath = mountPathRaw.(string)
+}
+options := secret.Data["options"]
+if options == nil {
+return mountPath, 1, nil
+}
+versionRaw := options.(map[string]interface{})["version"]
+if versionRaw == nil {
+return mountPath, 1, nil
+}
+version := versionRaw.(string)
+switch version {
+case "", "1":
+return mountPath, 1, nil
+case "2":
+return mountPath, 2, nil
+}
+
+return mountPath, 1, nil
+}
